Move some AI functions into new EspionageAI module
